% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{hadis_influence}
\alias{hadis_influence}
\title{hadis_influence's Influence Measure Calculation}
\usage{
hadis_influence(model)
}
\arguments{
\item{model}{An object of class \code{lm} that is the linear model for which Cook's distance is going to be calculated.}
}
\value{
A numerical vector that contains hadis_influence's influence measure of each observation in the model.
}
\description{
This function calculates the hadis_influence's influence measure that considers that influential observations may occur in the response variable, the predictors, or in both.
}
\examples{
data <- data.frame(
y <- c(4.2, 5.1, 7.3, 6.5, 8.0, 6.7, 10),
x1 <- c(2.1, 3.5, 5.1, 4.3, 6.0, 5.5, 7)
)
model <- lm(y ~ x1, data <- data)
hadis_influences <- hadis_influence(model)
}
