% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeasures.R
\name{plot_infmeasures}
\alias{plot_infmeasures}
\title{Plot Influence Measures for a Linear Model}
\usage{
plot_infmeasures(model, data, measure = "cooks", threshold = NULL)
}
\arguments{
\item{model}{An object of class \code{lm} that represents the linear model used to calculate the measures of influence.}

\item{data}{A dataframe that contains the data used to fit the linear model.}

\item{measure}{A character string specifying the measure of influence to be plotted. Options are \code{"cooks_distance"}, \code{"diffits"}, or \code{"hadis_influence"}. If not specified, \code{"cooks_distance"} will be used as the default.}

\item{threshold}{A numeric value specifying the threshold above which an observation is considered influential. If \code{NULL}, the function uses the default threshold based on the selected measure of influence.}
}
\value{
A ggplot object displaying the influence measures with outliers highlighted. The plot includes vertical lines representing the influence values for each observation and a horizontal line indicating the threshold.
}
\description{
This function plots various influence measures (Cook's Distance, DFFITS, Hadi's Influence Measure) for a given linear model.
}
\examples{
#' data <- data.frame(
y <- c(4.2, 5.1, 7.3, 6.5, 8.0, 6.7, 10),
x1 <- c(2.1, 3.5, 5.1, 4.3, 6.0, 5.5, 7)
)
model <- lm(y ~ x1, data <- data)
plot_infmeasures(model, mtcars, measure = "cooks_distance")
plot_infmeasures(model, mtcars, measure = "diffits", threshold = 0.5)
plot_infmeasures(model, mtcars, measure = "hadis_influence")
}
