% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{valInputs}
\alias{valInputs}
\title{Validate Inputs and Outputs}
\usage{
valInputs(data, model)
}
\arguments{
\item{data}{A dataframe  that contains the data  used to fit the linear model.}

\item{model}{An object of class \code{lm} that is the linear model used to calculate the measures of influence (cooks_distance, diffits, hadis_influence).}
}
\value{
The function returns \code{TRUE} if all validations pass. If any validation fails, the function stops execution and returns an error message.
}
\description{
This function validates the inputs to ensure that the provided data and model are appropriate for calculating influence measures.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
valInputs(mtcars, model)
}
