% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{cooks_distance}
\alias{cooks_distance}
\title{Cook's Distance Measure Calculation}
\usage{
cooks_distance(model)
}
\arguments{
\item{model}{An object of class \code{lm} that is the linear model for which Cook's distance is going to be calculated.}
}
\value{
A numerical vector that contains the Cook's distance of each observation in the model.
}
\description{
This function calculates Cook's distance, which is a measure of influence used to assess the impact that an observation has on the estimates of a least squares regression model.
}
\examples{
data <- data.frame(
y <- c(4.2, 5.1, 7.3, 6.5, 8.0, 6.7, 10),
x1 <- c(2.1, 3.5, 5.1, 4.3, 6.0, 5.5, 7)
)
model <- lm(y ~ x1, data <- data)
cooks_d <- cooks.distance(model)
}
