% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measuresCalculator.R
\name{diffits}
\alias{diffits}
\title{Difference in fits Measure Calculation}
\usage{
diffits(model)
}
\arguments{
\item{model}{An object of class \code{lm} that is the linear model for which Cooks distance is going to be calculated.}
}
\value{
A numerical vector that contains the difference in fits of each observation in the model.
}
\description{
This function calculates the difference in fits which is a metric used to identify influential data points by measuring how much the predicted value changes, in terms of standard deviations, when a specific observation is removed from the model.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
dfits <- diffits(model)
}
