% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measuresCalculator.R
\name{hadis_influence}
\alias{hadis_influence}
\title{Hadis Influence Measure Calculation}
\usage{
hadis_influence(model)
}
\arguments{
\item{model}{An object of class \code{lm} that is the linear model for which Cooks distance is going to be calculated.}
}
\value{
A numerical vector that contains Hadis influence measure of each observation in the model.
}
\description{
This function calculates the Hadis influence measure that considers that influential observations may occur in the response variable, the predictors, or in both.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
hadis_influences <- hadis_influence(model)
}
