% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{hadis_influence}
\alias{hadis_influence}
\title{hadis_influence's Influence Measure Calculation}
\usage{
hadis_influence(model)
}
\arguments{
\item{model}{An object of class \code{lm} that is the linear model for which Cook's distance is going to be calculated.}
}
\value{
A numerical vector that contains hadis_influence's influence measure of each observation in the model.
}
\description{
This function calculates the hadis_influence's influence measure that considers that influential observations may occur in the response variable, the predictors, or in both.
}
\examples{
model <- lm(mpg ~ disp + hp + wt + qsec, data = mtcars)
hadis_influences <- hadis_influence(model)
}
