% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{valInputs}
\alias{valInputs}
\title{Validate Inputs and Outputs}
\usage{
valInputs(data, model)
}
\arguments{
\item{data}{A dataframe  that contains the data  used to fit the linear model.}

\item{model}{An object of class \code{lm} that is the linear model used to calculate the measures of influence (cooks_distance, diffits, hadis_influence).}
}
\value{
The function returns \code{TRUE} if all validations pass. If any validation fails, the function stops execution and returns an error message.
}
\description{
This function validates the inputs to ensure that the provided data and model are appropriate for calculating influence measures.
}
\examples{
data <- data.frame(
y <- c(4.2, 5.1, 7.3, 6.5, 8.0, 6.7, 10),
x1 <- c(2.1, 3.5, 5.1, 4.3, 6.0, 5.5, 7)
)
model <- lm(y ~ x1, data <- data)

valInputs(data, model)
}
